had to make changes to sql file and then reload changes from echo to print
Not sure what happened between the last time I ran the program and now but our tables from our database was deleted off the local server. Had to change the name of the database from events to orlando-events and then reload it locally and on the server. Then reloaded the changes from echos to prints