scroll buttons only appear when viable
if to the left, only right are vis, vice versa.

this should really be a for loop so we can factor in different row counts, cards inside each row, and if where we are in scroll in case we have more than 8 items in row.

this is a concern for later! lol